day13/test_lib.py
import pytest
from lib import Seating, part1, part2


def test_lib():
    seating = Seating([example[0]])
    assert len(seating.guests) == 1
    assert "Alice" in seating.guests
    assert len(seating.happiness) == 1
    assert "Alice", "Bob" in seating.happiness
    assert seating.happiness["Alice", "Bob"] == 54


example = [
    "Alice would gain 54 happiness units by sitting next to Bob.",
    "Alice would lose 79 happiness units by sitting next to Carol.",
    "Alice would lose 2 happiness units by sitting next to David.",
    "Bob would gain 83 happiness units by sitting next to Alice.",
    "Bob would lose 7 happiness units by sitting next to Carol.",
    "Bob would lose 63 happiness units by sitting next to David.",
    "Carol would lose 62 happiness units by sitting next to Alice.",
    "Carol would gain 60 happiness units by sitting next to Bob.",
    "Carol would gain 55 happiness units by sitting next to David.",
    "David would gain 46 happiness units by sitting next to Alice.",
    "David would lose 7 happiness units by sitting next to Bob.",
    "David would gain 41 happiness units by sitting next to Carol.",
]